Finally made both kafka (confluent and kafka-python work).
The issue with confluent kafka is that you need to instatiate the subscription and the consumer instance together.

The issue with kafka-python API is that sometimes if you start consumer after producer had enqueued messages into kafka, the consumer is not able to dequeue those messages. You do not see this issue when you use confluent kafka.

Also, after dequeuing a message via confluent kafka or kafka-python API, the consumer is disconnected and subscription handle is terminated and the thread or the instance invoking this consumer API is disconnected from kafka. The reason why this is done is because we want kafka to start assigning job to the other consumer instances who have subscribed to the same topic and belong to the same group.

This is how the load balancing of jobs gets done in such a way that the jobs are equally distributed across all consumers that belong to the same topic.

Now, the next step is to implement auto scaling algorithm to scale consumers based upon the total enqueue count (vs) total dequeue count read from redis DB.